# OceanOfGameObjects.Model/Models/GameObjectModel.cs
using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace OceanOfGameObjects.Model
{
    public class GameObjectModel : GameObjectModelInsert
    {
        public int Gameid { get; set; }
    }
}